refactor: split shared dependency usage helpers
